refactor: MemochatWebView 컴포넌트로 분리 (src 폴더 생성)
